% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mornington.R
\name{mornington_crescent}
\alias{mornington_crescent}
\title{Play Mornington Crescent versus the console.}
\usage{
mornington_crescent()
}
\description{
Play Mornington Crescent versus the console.
}
\examples{
  mornington_crescent()
}
